优化了动态模板的命名

+++ fox-edge-server-kernel/fox-edge-server-kernel-system/fox-edge-server-system-repository/src/main/java/cn/foxtech/kernel/system/repository/service/RepoLocalCompService.java
@@ -126,7 +126,7 @@ public class RepoLocalCompService {
         this.entityManageService.deleteEntity(compId, RepoCompEntity.class);
     }
 
-    public RepoCompEntity getRepoCompEntity(Integer compId) {
+    public RepoCompEntity getRepoCompEntity(Long compId) {
         return this.entityManageService.getEntity(Long.parseLong(compId.toString()), RepoCompEntity.class);
     }
